Move tenant code to ElvTenant.js
